Bash Script to run the python file
Created the bash script to run the commands on the python file. Per the professor the size and ways are hard coded. However, his feedback to me regarding the combinations I used was "find the best combination that allows you to compare all the trace files." So, some of these command lines will have to change. But for now ...